fixed tile buffer type (int64 -> float64). #107

+++ cmd/tegola/cmd/cache.go
@@ -206,7 +206,7 @@ var cacheCmd = &cobra.Command{
 
 						//	set tile buffer if it was configured by the user
 						if conf.TileBuffer > 0 {
-							mt.Tile.Buffer = conf.TileBuffer
+							mt.Tile.Buffer = float64(conf.TileBuffer)
 						}
 
 						//	seed the tile
